document_ai_app/requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# OCR Dependencies
pytesseract>=0.3.13
Pillow>=10.0.0
paddlepaddle>=3.0.0
paddleocr>=3.4.0

# PDF Processing
PyMuPDF>=1.23.0

# AI and LLM Dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0

# Enhanced OCR dependencies
docling>=1.0.0
docling-ai>=1.0.0
paddleocr>=2.7.0
landingai-ade>=0.1.0

# Frontend Dependencies
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.25.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
